refactor: split RunDetailContent into section components (#110)
